Feature 3302 added: multithreading for Perl jobs. We don't use threads at all in fact. It's just the way it was first implemented in Java. As recommended by Perl (and Unix) gurus, we're using processes, we fork.
Modification: graphical statistics were redesigned for Perl. We don't use a
specific thread any longer. The thread model was not working with
"multithreading" new mode. The decision to send a message in the socket was
moved to component_part_header.perljet : if the last message was send more
than 2 seconds ago, we send a message. The new model is far faster than the
previous one. On my bench job (1M lines), the old model was decreasing
performances by 50% while the new model only decreases performances by 5%.

Modification: after an iterate link, the graphical statistics are
reinitialized.

git-svn-id: http://talendforge.org/svn/tos/trunk@12159 f6f1c999-d317-4740-80b0-e6d1abc6f99e