Bug TDI-20080 fixed: Move a job into Referential view does'nt work or is not smooth.
Root cause for moving job is not smooth as before:

Before really move the job, the code checks whether the job is locked or not, this check spends some time because it will not only check the local user, but also the remote users. And current there is a check every time dragging a job to a new target(including folders and jobs), so it will wait at every job and folder which the job might pass when dragging it to another folder.

Fix Solution: 
Do not check the lock status when draging a job until the draging is finished, before really moving the job, check the lock status, if it's locked,  prompt a message to tell users that the job is locked

Root cause for moving job doesn't work:

When the mouse moves over the drop target, if it's a valid drop, RepositoryDropAdapter records the event.detail as operation(DND.DROP_MOVE, DND.DROP_COPY, DND.DROP_DEFAULT), and when drop the object, we perform the copying or moving operation according to the recorded operation(DND.DROP_MOVE, DND.DROP_COPY,DROP_DEFAULT).  Sometimes when moving  over the drop target, the operation is recorded as DND.DROP_DEFAULT, but when dropping the object, the event.detail == DND.DROP_MOVE. This could be a bug of eclipse, as the recorded operation does not match the actually operation.

Fix Solution: 
Record the drop event, and distinguish the operation type by the drop event.

git-svn-id: http://talendforge.org/svn/tos/branches/branch-5_0@79974 f6f1c999-d317-4740-80b0-e6d1abc6f99e